* FIXED: fixed some compile warnings, please be careful guys. * FIXED: some PvP flagging stuff (mostly spell and areachange related). * FIXED: Applied Jackpoz's improved spellreflect patch. Fixes the related crash. Thanks for contributing! http://arcemu.org/forums/index.php?showtopic=17943

+++ src/arcemu-world/MailSystem.cpp
@@ -90,7 +90,7 @@ WorldPacket * Mailbox::BuildMailboxListingPacket()
 	}
 
 	data->put<uint32>(0, realcount); 
-	data->put<uint8>(4, count); 
+	data->put<uint8>(4, static_cast< uint8 >( count )); 
 
 	// do cleanup on request mail
 	CleanupExpiredMessages();
